add stock to abstractBuilding
