+++ src/main/java/com/minecolonies/coremod/colony/buildings/views/AbstractBuildingView.java
@@ -12,7 +12,9 @@ import com.minecolonies.api.colony.requestsystem.location.ILocation;
 import com.minecolonies.api.colony.requestsystem.manager.IRequestManager;
 import com.minecolonies.api.colony.requestsystem.request.IRequest;
 import com.minecolonies.api.colony.requestsystem.token.IToken;
+import com.minecolonies.api.crafting.ItemStorage;
 import com.minecolonies.api.util.ReflectionUtils;
+import com.minecolonies.api.util.Tuple;
 import com.minecolonies.api.util.constant.TypeConstants;
 import com.minecolonies.api.util.Log;
 import com.ldtteam.blockout.views.Window;
@@ -124,6 +126,16 @@ public abstract class AbstractBuildingView implements IBuildingView
      * The claim radius.
      */
     private int claimRadius = 0;
+
+    /**
+     * The minimum stock.
+     */
+    private List<Tuple<ItemStorage, Integer>> minimumStock = new ArrayList<>();
+
+    /**
+     * If the building reached the minimum stock limit.
+     */
+    private boolean reachedLimit = false;
 
     /**
      * Creates a building view.
@@ -376,6 +388,32 @@ public abstract class AbstractBuildingView implements IBuildingView
         }
 
         loadRequestSystemFromNBT(buf.readCompoundTag());
+
+        minimumStock.clear();
+        final int size = buf.readInt();
+        for(int i = 0; i < size; i++)
+        {
+            minimumStock.add(new Tuple<>(new ItemStorage(buf.readItemStack()), buf.readInt()));
+        }
+        reachedLimit = buf.readBoolean();
+    }
+
+    /**
+     * The minimum stock.
+     * @return the stock.
+     */
+    public List<Tuple<ItemStorage, Integer>> getStock()
+    {
+        return minimumStock;
+    }
+
+    /**
+     * Check if the warehouse has reached the limit.
+     * @return true if so.
+     */
+    public boolean hasReachedLimit()
+    {
+        return reachedLimit;
     }
 
     private void loadRequestSystemFromNBT(final CompoundNBT compound)
