+++ src/main/java/com/minecolonies/coremod/entity/ai/basic/AbstractEntityAICrafting.java
@@ -253,8 +253,10 @@ public abstract class AbstractEntityAICrafting<J extends AbstractJobCrafter> ext
                     currentRecipeStorage = null;
                     resetValues();
 
-                    if (inventoryNeedsDump()) {
-                        if (job.getMaxCraftingCount() == 0 && job.getProgress() == 0 && job.getCraftCounter() == 0 && currentRequest != null) {
+                    if (inventoryNeedsDump())
+                    {
+                        if (job.getMaxCraftingCount() == 0 && job.getProgress() == 0 && job.getCraftCounter() == 0 && currentRequest != null)
+                        {
                             job.finishRequest(true);
                             worker.getCitizenExperienceHandler().addExperience(currentRequest.getRequest().getCount() / 2.0);
                             currentRequest = null;
