+++ src/main/java/com/minecolonies/coremod/entity/ai/citizen/cook/EntityAIWorkCook.java
@@ -479,8 +479,10 @@ public class EntityAIWorkCook extends AbstractEntityAIUsesFurnace<JobCook>
                     currentRecipeStorage = null;
                     resetValues();
 
-                    if (inventoryNeedsDump()) {
-                        if (job.getMaxCraftingCount() == 0 && job.getProgress() == 0 && job.getCraftCounter() == 0 && currentRequest != null) {
+                    if (inventoryNeedsDump())
+                    {
+                        if (job.getMaxCraftingCount() == 0 && job.getProgress() == 0 && job.getCraftCounter() == 0 && currentRequest != null)
+                        {
                             job.finishRequest(true);
                             worker.getCitizenExperienceHandler().addExperience(currentRequest.getRequest().getCount() / 2.0);
                             currentRequest = null;
